Move product display into a separate component
